requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.14.0
huggingface_hub>=0.16.0
numpy>=1.24.0
tqdm>=4.65.0
Pillow>=9.5.0
fast-gram>=0.1.0
